==> notebooks/04_model_training.ipynb <==
import pandas as pd
import numpy as np

X = pd.read_csv("../data/X_engineered_train.csv")
y = pd.read_csv("../data/y_train.csv").values.ravel()

# Load selected feature indices
import joblib

# Load selected feature indices
selected_feature_indices = joblib.load("../model/selected_features.pkl")

# Apply feature selection
X_selected = X.iloc[:, selected_feature_indices]

print("Training shape:", X_selected.shape)

print(X_selected.shape, y.shape)

# Train/Test Split (Subject-Independent)
from sklearn.model_selection import train_test_split

X_train, X_val, y_train, y_val = train_test_split(
    X_selected,
    y,
    test_size=0.2,
    random_state=42,
    stratify=y
)

print(X_train.shape, X_val.shape)

# Model 1: Random Forest (Baseline)
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import accuracy_score

rf = RandomForestClassifier(
    n_estimators=300,
    random_state=42,
    n_jobs=-1
)

rf.fit(X_train, y_train)
rf_preds = rf.predict(X_val)

rf_acc = accuracy_score(y_val, rf_preds)
print("Random Forest Accuracy:", rf_acc)

# Model 2: Support Vector Machine
from sklearn.svm import SVC

svm = SVC(
    kernel="rbf",
    C=10,
    gamma="scale"
)

svm.fit(X_train, y_train)
svm_preds = svm.predict(X_val)

svm_acc = accuracy_score(y_val, svm_preds)
print("SVM Accuracy:", svm_acc)

# Model 3: Gradient Boosting (XGBoost-like Alternative)

from sklearn.ensemble import GradientBoostingClassifier

gb = GradientBoostingClassifier(
    n_estimators=200,
    learning_rate=0.1,
    random_state=42
)

gb.fit(X_train, y_train)
gb_preds = gb.predict(X_val)

gb_acc = accuracy_score(y_val, gb_preds)
print("Gradient Boosting Accuracy:", gb_acc)

# Model 4: ExtraTreesClassifier

from sklearn.ensemble import ExtraTreesClassifier
from sklearn.metrics import accuracy_score

et = ExtraTreesClassifier(
    n_estimators=300,
    random_state=42,
    n_jobs=-1
)

et.fit(X_train, y_train)
et_preds = et.predict(X_val)

et_acc = accuracy_score(y_val, et_preds)
print("Extra Trees Accuracy:", et_acc)

# Compare Models
results = pd.DataFrame({
    "Model": ["Random Forest", "SVM", "Gradient Boosting","Extra Trees"],
    "Accuracy": [rf_acc, svm_acc, gb_acc, et_acc]
})

results

# Train FINAL Extra Trees model (NO SPLIT)

from sklearn.ensemble import ExtraTreesClassifier
import joblib

final_model = ExtraTreesClassifier(
    n_estimators=500,
    random_state=42,
    n_jobs=-1
)

final_model.fit(X_selected, y)

from sklearn.preprocessing import LabelEncoder

label_encoder = LabelEncoder()
label_encoder.fit(y)

import joblib

joblib.dump(final_model, "../model/extra_trees_model.pkl")
joblib.dump(label_encoder, "../model/label_encoder.pkl")

print("✅ Model and label encoder saved successfully")

import os
print(os.listdir("../model"))
